ui/requirements.txt
# Core dependencies
streamlit>=1.27.0
pandas>=2.1.1
plotly>=5.17.0
httpx>=0.25.0
python-dotenv>=1.0.0

# Authentication
streamlit-authenticator>=0.2.3
pyjwt>=2.8.0

# Data handling
xlsxwriter>=3.1.4
openpyxl>=3.1.2

# Visualization
altair>=5.1.2
matplotlib>=3.8.0
seaborn>=0.13.0
streamlit-echarts>=0.4.0

# UI enhancement
streamlit-extras>=0.3.4
streamlit-card>=0.0.61
streamlit-toggle-switch>=1.0.2
streamlit-option-menu>=0.3.6

# Development
black>=23.9.1
isort>=5.12.0
flake8>=6.1.0